user/dnscrypt-proxy: update to 2.1.12
The default configuration file, dnscrypt-proxy.toml, was removed in
favor of patching the configuration file in builddir. A helper python
script, generate-domains-blocklist.py, along with other default
configuration files were added for a better out of the box experience.

The dnscrypt-proxy service file was updated to the latest dinit service
file syntax.

We can't add run-as = _dnscrypt in the dinit service file because
dnscrypt-proxy fails to bind to the port 53 in that case.

closes #4089